* Move PulseSpilloverNoiseClass in separate script * Adjust noise profile access in HPO * Remove ununsed function in CAQEnv
